refactoring : refactored profile navigation.
