+++ src/modules/navigation/hooks/ui/useProfileNavbar.jsx
